requirements.txt
google-generativeai
python-dotenv
tqdm
Pillow
PyMuPDF